feat: retry with reverse complement when seed matching fails
Adds flag `--retry-reverse-complement` which enables additional attempt of seed matching when initial attempt fails. The second attempt is performed on reverse-complemented sequence.

As a consequence, the output alignment, peptides and analysis results correspond to this modified sequence and not to the original.

This functionality is opt-in and the default behavior is to skip sequence with a warning.