fix(cli): prevent truncation of  components if basename contains dots
Resolves: https://github.com/nextstrain/nextclade/issues/907

This rolls an in-house version of `add_extension()` function which always adds an extension to a `PathBuf`.  This is different from `PathBuf::with_extension()` which may replace or add extension depending on what the path is.

This solves a problem with basenames containing dots, as described in the issue: `PathBuf::with_extension()` thought that they are extensions and replaced the last one. But we always want to add, not replace.