feat(cli): add 'warnings' and 'failedGenes' columns to results csv/tsv
This adds new columns to main nextclade results CSV and TSV outputs: `warnings` and `failedGenes`, which include any warnings emitted for a sequence as well as a list of genes that failed translation.

Now all columns of the "errors.csv" file are also in the CSV and TSV results files.