feat: add "download dataset and run" shortcut"
This adds `--dataset-name` (`-d`) to `run` command, which allows to download a dataset with default parameters and run with it immediately, in one command.

For example this command.

```bash
nextclade run --output-all=out --dataset-name=sars-cov-2 sequences.fasta
```

will download the latest default sars-cov-2 into memory and will run analysis with these dataset files.

This is a convenience shortcut for the usual combination of `dataset get` + `run`. The dataset is not persisted on disk and downloaded on every run.