fix: use indices to identify sequences uniquely
Nextclade is using sequence names to uniquely identify sequences. However, sequence names come from user inputs and cannot be trusted to be unique. Neither it seems there is a consensus on uniqueness of sequence names in the bioinformatics community as a whole.

This causes various problems where sequence names are used as identifiers, and when for example there are multiple sequences with the same name. In particular, when storing analysis results, they are effectively stores in an associative container, where sequence name acts as a key. This leads to newer results overwriting older results as they arrive during analysis. Additionally, some of the HTML `id` properties used sequence names to add uniqueness. This was leading to incorrect HTML being produces, with multiple elements having the same `id` property.

In this PR I:
 - change internal storage to use sequence indices in the input file(s) as keys
 - add sequence index into HTML `id`s
 - display "Sequence index" in places where only sequence name was displayed previously

This should ensure correct handling of duplicated names.

This affects only web application. In the algorithmic and CLI parts, sequence names are not used - results are stored in the form of an array, and no HTML is involved.