fix: consider different file encodings
This attempts to mitigate crash in web when the fasta file is not UTF-8 or ASCII encoded.

Here I am trying to deduce a list of encodings the file resembles and then attempting to decode the contents of the file from these encodings to a JS string, instead of simply using the `FileReader`'s `.toString()` function.