feat(web): dynamically set dataset server root using a URL param
Allows to dynamically set dataset server root URL in Nextclade Web on runtime.

It is sufficient to add URL parameter `dataset-server` with the value containing URL to the root of the dataset file server (containing `index_v2.json`), and Nextclade Web will fetch dataset index and all dataset files from this server instead of the default:

```
?dataset-server=http://example.com
```

The localhost URLs should also work:
```
?dataset-server=http://localhost:27722
```

This is equivalent to `--server` flag in Nextclade CLI. 

This should facilitate testing of custom datasets in Nextclade Web.