feat(web): reduce default number of threads to speedup initialization
Despite "thread" (actually webworker) initialization is supposed to be concurrent, there seem to be a clear trend of increased initialization time as number of threads increased. This is likely due to data transfer overhead between workers.

Here I reduce number of threads that can be initialized with default settings.

But these setting can always be altered in the "Settings" dialog.