refactor: silence ts errors due to defects in i18n
Again, due to downgrade in commit `3604038`, the older versions of i18n causing troubles with Typescript types.

Here I just silence the 2 errors, because the typings for `I18nextProvider` are obviously bogus. This is fixed in later versions, so when Auspice upgrades the i18n packages, these `ts-ignore` statements can be removed.