chore(ci): publish to bioconda
Adds a script and a GHA workflow to publish new version on bioconda. 

The workflow is normally to be triggered by a GitHub release (but can also be triggered manually or through API). It downloads the binaries, gets their version,  commits and pushes updated `meta.yaml` to https://github.com/nextstrain/bioconda-recipes and then creates a PR from there to https://github.com/bioconda/bioconda-recipes/branches.