fix: set default weights in private mutations qc check to 1
This fixes the bug when the QC score is 0 when the following QC fields are missing from `qc.json`:
```
    .privateMutations.weightLabeledSubstitutions
    .privateMutations.weightReversionSubstitutions
    .privateMutations.weightUnlabeledSubstitutions
```
In this case Nextclade assumed value 0, and ending up with calculating QC score of 0. Not all datasets were adjusted, so some datasets don't have these fields.

Here I set the default of these weights to 1.0 if they are missing, such that QC have some sensible values again.