fix: ensure `--output-all`, `--output-tsv`, `--output-csv` can be used together again
Followup of https://github.com/nextstrain/nextclade/pull/1081

Turns out setting `group` for related arguments in `clap` makes them exclusive, so for example `--output-all` and `--output-tsv` could not be used together after the changes in https://github.com/nextstrain/nextclade/pull/1081.

In order to fix that I remove the `group` and check the presence of one of `--output-all`, `--output-csv`, `--output-tsv` manually. This is the constraint required for `--output-columns-selection`. I tried to express with the `group`, but that did had the aforementioned unwanted side effect, breaking bioconda tests: https://github.com/bioconda/bioconda-recipes/pull/38990