chore: `cargo update` within Cargo.toml constraints
`cargo update` updates dependencies within the constrains of Cargo.toml,
i.e. not changing the Cargo.toml

Cargo.toml is changed only for bio-types, which needs to be the same version
as that depended on by bio, because bio uses bio-types Strand in its public
API, without properly useing `pub use bio-types` which, see
https://github.com/rust-bio/rust-bio/issues/516

The solution here would be for `bio` to `pub use` then we can get rid of the
`bio-types` dependency.