fix(web): ensure correct width of gene map
Resolves https://github.com/nextstrain/nextclade/issues/1084

The gene map (genome annotation) schema is not aligned with sequence views (it is narrower).

This was caused by extra width added to the gene map name column to the left of the schema. There are clade-like columns in the inputs that are hidden from web app. However, they were still counted when calculating gene map width.

Here I filter out the hidden clade-like entries before calculating width. This should resolve the issue.