feat(web): always show action buttons on results page
ON smaller screens the "Download", "Tree" and other buttons were overflowing, so that they are not visible by default and horizontal scrolling is required to see them.

Here I change the layout such that the panel with buttons does not overflow along with table and so the buttons are always visible. Table is still scrollable.