feat: improve alignment errors
Resolves https://github.com/nextstrain/nextclade/issues/554


As concluded in the issue, here I just append another sentence to the text of alignment errors, to add some more context and actionable information to the user:

```
This is likely due to a low quality of the provided sequence, or due to using incorrect reference sequence.
```

As before, these messages will appear as errors for nuc alignment and as per-gene warnings for aa alignment.

Additionally I modified the message for "no seed matches" error. It is triggering when there is less than 2 seed matches, so  "no seed matches: 1 seed match" seemed a bit confusing. Now it says "not enough seed matches".