chore: remove `DOCKER_BUILDKIT` var
Resolves https://github.com/nextstrain/nextclade/issues/1107

(or rather attempts to)

We are using `docker buildx` command, which AFAIU implies `DOCKER_BUILDKIT=1`.  
But someone on the internet* suggested on the internet that removing it or setting `DOCKER_BUILDKIT=0` may resolve the errors we are having. I'd like to keep using buildkit though, so don't know if any of it is a good idea. But let's give it a try on a branch.

 \* - https://github.com/docker/buildx/issues/415, https://stackoverflow.com/questions/64221861/an-error-failed-to-solve-with-frontend-dockerfile-v0