feat: use placement priors as tie breaker
This feature improves the placement of sequences that
could be added to multiple nodes on the tree
with the same distance

As tie-braker we use "priors", a number that is present on every tree node

That number represent an a priori likelihood
that sequences attach to that node

The calculation details of this prior could change. All that matters for now
is the ordering of priors

The numerical value has no purpose (yet)
In the future, we intend to use the priors
to display placement confidence information.

This feature should be backwards compatible. Trees without priors should work just as before.

This commit is a placement-only subset of PR #1114