feat: only emit nearest node info if a CLI flag is set
Introduces a CLI flag `--include-nearest-node-info` and the correponding internal parameter for Nextclade runs. Skips the `.nearestNodes` field in JSON/NDJSON unless this flag is present.

This parameter is always `false` in web, so this data is never included into the outputs of the web app. This also helps to slightly reduce the amount of data passed between webworkers.