fix: ensure phenotype values are attached to new tree nodes
This adds custom phenotype values (such as `ace2_binding` and `"immune_escape"`) to the newly placed tree nodes.

They were previously missing, even though the technically similar clade-like node attributes were set. Which makes it somewhat asymmetric. 

I think it was simply an omission when the phenotype values were added.

Here I add the `phenotype_values` to the node similarly to how the `custom_node_attributes` are added.

This allow to see the values and colorings for phenotype values on the tree page.