fix: incorrect 0 coercion
The `pos` of  `0` caused to show mutation badges with position 0, while they are supposed to be 1-indexed. This was due to incorrect conversion of `0` to `false`.

Here I fix the problem by using the explicit comparison to `undefined` and `null` (using the helper function `isNil()` from `lodash`), which avoids implicit conversions.