fix(web): sort empty values
Resolves https://github.com/nextstrain/nextclade/issues/1143

The `undefined` values are not comparable to strings, so Cornelius felt that in presence of empty values in custom clade-like columns the table sort misbehaves.

Here I treat the `undefined` values in custom clade-like columns and for phenotype columns as empty strings before sorting them. This way they can be sorted lexicographically along with other string values.

This may or may not what we want, but let's give it a try.