feat(web): notify user on Nextclade Web updates
Related to: https://github.com/nextstrain/nextclade/issues/1139 https://github.com/nextstrain/nextclade/pull/1147

This PR introduces a notification toast which is displayed to the user when a new version of Nextclade Web is released. User can:
 - accept the update - this simply hard-reloads the page and, thus, loads the new version of the app
 - dismiss the notification - this hides the notification until the next version is available
 - read changelog - this opens changelog document on GitHub (the new changelog markdown is not yet available in the old app, so we cannot render it nicely and we link to GitHub instead)


### Implementation

The version info is fetched from the server as a field in the new file, `app.json` which is created by webpack and placed into the webroot. The version in this file (always latest available on the server) is compared with the version of the app currently running.

If user dismisses the toast, the last displayed version is remembered in local storage. This version is taken into account on next visit and user is not notified about this particular version anymore.


### Screenshot

![Update notification toast](https://user-images.githubusercontent.com/9403403/236952750-659c843d-497b-42fe-ab73-9591965ce518.png)