fix(web): prevent main page from blinking when selecting dataset
The main page blinks when selecting dataset, because at this point background update fetches are running, triggering React Suspense.

In this case we don't need to block the UI with suspense though, because these are background checks and their results are deferred to later time.

So let's disable suspense for the background checks for: 
 - dataset index update
 - current dataset version update
 - software update