fix(web): correct type of nua/aa sequences
We are converting `Nuc[]` and `Aa[]` to `String` during serialization (and back during deserialization) on Rust side, so the web app sees only `string`. Here I add an annotation to struct fields, such that JSON schema and Typescript typings are generating with the correct `string` type, rather than `Nuc[]` and `Aa[]`.