perf: avoid extra copies of genes and cdses
The structs for translated genes and CDSes also contain copies (the .clone() is called on construction) of their corresponding gene and CDS annotations (from gene map). This is unnecessary in many cases - only the names and strands are accessed. Let's try to shave off a few microseconds of runtime by removing these clones.