refactor: introduce safe tagged position and range types
Tagged position and range types for different coordinate spaces.

Each coordinate space has a dummy marker (tag) struct implementing a coordinate space trait. A concrete position type and its sibling range type use these markers to distinguish themselves from positions and ranges in other spaces. This way the positions and ranges in different spaces have different Rust types and they cannot be accidentally mixed up.

This prevents, for example, adding a position in alignment coordinates to the position in the reference coordinates, which is always a bug. Similarly, you cannot passing a range in reference to a function  expecting a range in alignment.