fix: frame across cds segments that are not multiples of 3
the previous implementation had a loop counter that counted along
the cds segment starting at 0 for each segment and was used to
calculate frame. But segments don't have to be multiples of 3
and the frame persists across segments. this moves the cds position
counter outside of the segment loop and increments it across the entire
cds.

added a test.