fix: remove `strand` field from all features except CDS segments
GFF format allows all genetic features to have a strand information associated to them. 

But in context of Nextclade, most features have nothing to do with strands. Strand is only relevant when extracting a range from a sequence. Only 'CDS segments' are being extracted in Nextclade, so they are the only ones to carry this property. It is not even clear what is a strand for a gene, for the whole CDS or for a peptide is and how to deduce it. So we chose to remove it from anywhere else other than 'CDS segments'.

Behavior had to be adjusted in 2 places where the CDS were being extracted. Previously the whole CDS was being reverse-complemented in case all segments were on reverse strand. Now each segment is being reverse-complemented separately, before concatenation.

This allows to support CDSes consisting from segments on different strands.