fix(web): prevent browser from hanging when there's many Ns
It seems that because the `missing.range` was not destructured correctly, some of the "Missing" markers were having duplicated same HTML ids. The exact mechanism is unknown (implemented in a library component), but this lead to tooltip component hanging the browser for a while when rendering sequences with many Ns, even when the tooltip is hidden.

TODO:
 - extract tooltip shown/hidden logic into a separate component and memoize 
 - make sure tooltips are not added to the DOM until hovered
 - there should be a way to anchor tooltips without the unreliable ad-hoc HTML ids (by using DOM nesting)