perf(web): send minified json between wasm and js
Currently JSON being sent from wasm to js is prettified. This means more bytes are being sent across language boundaries and might cause unnecessary slowdown.

Here I am adding a function to serialize to JSON without prettifying it. Let's see if it has any visible effect on performance.

However it also means that the downloadable JSON is also minified. Or it will need to be prettified on demand.