feat(web): reimplement gene selector in results table
Each entry in the dropdown menu is now denoted with its feature type.

The menu now shows a hierarchy of gene-CDS relationships if a gene contains more than one CDS. Otherwise it's a flat list as before.

Currently, only CDS entries can be selected, because we don't have anything to show for other types of features. But in theory, any entry can be selected if needed.


We can use this new widget to display arbitrary trees of genetic features. For example, we can nest  proteins under CDSes.

More details can be displayed, e.g. frame, phase, length and boundaries of each feature etc.