fix: ensure results are sorted in both cli and web app
The sorting of results before `graph_attach_new_nodes_in_place()` was previously only done in CLI code. The wasm module of the web app did not sort the results.  Let's move sorting inside `graph_attach_new_nodes_in_place()`, because this is a common functionality and a precondition for the function to work correctly.