fix: fine tuning when attaching to terminal node
When a sequence attaches to a terminal node, we introduce an aux
node above it. This often results in zero-length branches. When other
sequences attach to the same node, this previous fine-tuneing did not
move up to the parent across that zero length node and was stuck.
We now check for a zero-length branch when there is no shared mutations.