fix: flag to align gaps left
Historically, nextalign has aligned sequence around gaps on the right
when there were equivalent gap placements and we introduced a flag to
align left/right. The flag amounts to following a "match" or "gap" path
first and algorithmically can be dealt with a `<` vs `<=` comparison
of scores when evaluating options (pick first or last). Alas, this
switch was not correctly implemented and likely resulted in suboptimal
alignments for `--gap-alignment-side left`.