fix: don't consider ambiguous nucs fro reversions
There is a problem when searching for reversions.

Setup:

```
Nuc pos: 22770 (1-based)
Ref:  G
Node: A
Qry:  R
```

When searching for mutations, `R` is not considered a substitution. So when private mutations are searched it is not considered a mutated position. Next, when searching reversions, we see that the position is not mutated, but sequenced, so it is marked as a reversion.

This is wrong, because we cannot tell for whether reversion have taken place or not, due to ambiguity.

Here I exclude all query positions with non-ACGTN from consideration.