refactor: rename PrivateMutationsMinimal to BranchMutations
This object is used in tree building and collects all changes that
happen along a branch in the tree for nucleotides and amino acid changes.
It treats deletions and substitutions on the same footing as both need
to be considered simultaneously to maintain a consistent state of the
sequence along the tree.