fix: fasta files without trailing newline are read correctly
It was reported that when multiple fasta files is provided to Nextclade CLI, if some of the the files don't contain trailing newlines, the sequences aren't read correctly.

The issue comes from the fast that the `Concat` struct concatenates readers (e.g. files in this case) seamlessly, and last sequence of previous reader transitions immediately into the first sequence of the next reader. Then line parsing with `BufRead` cannot split these sequences, because there is no newline character between them.

To address this, here I modify `Concat` struct, adding an additional parameter `delimiter`. When provided, readers are concatenated with given delimiter bytes. IN fasta reader this allows to force insertion of a "\n" character between readers.

If files contain newlines and/or carriage returns, this might end up in multiple newline characters. But this is not a problem for fasta reader, because these get stripped away later.