fix: double mutation in codon when moving up the tree
when a samples shared all privated nuc mutations with a parent, we move
up the tree and subtract the shared mutations. However, in rare cases
when there are multiple mutations in a codon, not all amino acid mutations
are shared. hence we still need to handle left over mutations also
in the case where we just move up the tree, not only when we move down
the tree.

This change actually harmonized cases and reduced code duplication.