feat(web): add sequence markers for ambiguous nucleotides
Resolves: https://github.com/nextstrain/nextclade/issues/700

Add colored markers for ambiguous nucleotide to nucleotide sequence view.

I tried to pick colors to be pale (so that they resemble pale grey of missing nucs) and to be a mixture of colors of corresponding nucs, or to be inverse color in case of "not" characters (e.g. color of not-A is a pale inverse of color of A).

I made them half-height at the top by default, just like missing nucs. (Settings UI is currently broken, so it's impossible to change at the moment, sorry)

Tooltip text is as specified in the original issue.