feat: separate undeletions from reversion substitutions
In v2, we ignored undeletions (i.e. ref is gap, qry is ref)
Until now, in v3, we treat them as reversion substitutions
A single stretch 20 bps undeleted results in 20 reversion substitutions which results in unexpectedly large QC penalties (some penalty is ok but it's too large as is)

In this PR, I split undeletions from ordinary substitutions

Currently, nothing is done with the undeletions ranges, but we could surface them into TSV/web UI at a later point. We could also count each stretch/range towards QC as a single unit of a reversion (and analogously treat each private deletion stretch as a single private mutation).

We also don't seem to surface private deletions at the moment, so this could be done as a package.

Stripped down version of #1296 without aggregation of undeletions into ranges