fix: correctly filter CDSes in genome annotation
Genome annotation was not filtered correctly. The values from `--cds-selection` (previously `--genes`) were still used to filter genes, not CDSes.

Here I reimplement filtering, such that:

- first CDSes are filtered for each gene
- then empty genes are dropped

At the same time, I move gene storage from a map to vec, to prepare it for the better genome annotation validation (this allows me to avoid reimplementing the filtering twice - for map and then for vec, even though it's unrelated to this particular PR).