refactor: extract worker construction function back outside of the Pool
Instead of creating the `Pool` with hardcoded module path in the`new Worker()` call, let's try to move it back into a parameter. This way we could create worker pools for other workers, not just analysis worker (e.g. for autosuggestion worker or if we split the analysis worker into multiple workers)

In usual programming environments you'd just extract the path and make it a string parameter. But  if I recall correctly, this cannot be done here. These paths are resolved during build time by Webpack and the path does not exist on runtime - dynamic string cannot be passed to the `new Worker()` call. There is simply no `src/` directory exist in the deployed app.

The trick is to still have the hardcoded path in the `new Worker()` call, but wrap the call into a closure. This is how `threads.js` does it  - `spawn()` does not accept a `worker` instance, but a closure `() => new Worker()`, which it then calls as many times as needed to create as many workers as we request. Let's use the same trick and just pass the closure when creating the `Pool`. This way the `new Worker()` is still duplicated but at least we don't have to  copy-paste the entire `Pool` class to have a pool of different workers.