fix: handle leading newlines in fasta parser correctly
The previous version would throw an error if there's a leading newline in the inputs.

Because of the changes in `Concat` iterator in Nextclade v3, `FastaReader` now receives additional newlines, which `Concat` inserts between input readers (e.g. files or string streams). This was a necessary fix to ensure that there's a separation between two concatenated readers when the previous reader does not end with a newline. However, now if the first reader (or the only reader)  is empty or reads only newlines, then after `Concat` `FastaReader` receives only a `\n` characters, which it could not interpret and fails with an error. This is a change of behavior compared to Nextclade v2.

In the new version I make sure to skip the leading newlines. The danger of this version is the `loop` block. The termination of this loop entirely depends on the inputs, so it is important to find all corner cases to avoid infinite loop and to avoid terminating it too early.