fix(cli): crash when using `--verbosity` option
Resolves: https://github.com/nextstrain/nextclade/issues/1428

When using `--verbosity` option, Nextclade panicked with error:
```
The application panicked (crashed).
Message:  Mismatch between definition and access of `verbosity`. Could not downcast to log::LevelFilter, need to downcast to alloc::string::String
```
(in release builds the message can be different due to type mangling)

This is due to particularities of how `value_parser` works in `clap` v4, which is a replacement for `possible_values` functionality we used in `clap` v3, but which was removed in v4.

In Nextclade the `verbosity` filed is of type `log::LevelFilter`. It seems `clap` v4 can read a list of string values passed to `value_parser` and displays them correctly in the `--help` text, but when trying to actually pass one of these values to the `--verbosity` arg, `clap` would panic with a downcast error somewhere deep inside the library.

It's unclear to me why this is now suddenly needed, because `possible_values` worked just fine, but it seems that using an an iterator to the `clap`'s `PossibleValue`s constructed out of `LevelFilter` instead of strings work. I found this solution here: https://github.com/clap-rs/clap/discussions/4264. There are many more issues and discussions in clap repo about that. Maybe there's a better solution.

At the same time, I could not resist to clean up our `verbosity.rs` file, because most of the stuff we copied from `clap-verbosity-flag` crate and never actually use. I also use clap's `default_value` parameter to set a default value for the `verbosity` now, rather than setting a default manually on runtime.