fix(cli): don't use hardcoded filenames when reading dataset zip
We want to only read dataset files which are declared in the `pathogen.json`.Most of these files are optional and can be missing, so we should never rely on their presence or their particular names, wihout checking of what's declared in `pathogen.json`

However, in Nextclade CLI we've been reading hardcoded `tree.json` and `genome_annotation.gff3` whenever `--input-dataset` was given a dataset zip archive.

Let's made it to use `pathogen.json:files.treeJson` and `pathogen.json:files.genomeAnnotation`, just like it happens when using dataset in the form of directory, or like it happens in Nextclade Web.