fix(web): correctly select top dataset with autosuggest is disabled
Let's focus on the `getPromise()` calls followed by the `set(datasetCurrentAtom, ...)` call. This combo lags behind and does not yet receive the suggestion results we've just `set()` in the `onResult()` callback. With this config the "Suggest" button needs to be clicked twice for autoselection to take effect. And with current (quite complex) code organization I don't see any more obvious way to set the top dataset.

Let's use an additional boolean state which tells that the suggestion was triggered manually. We can then use it in the caller component, which would set. This guarantees ordering of operations and correct selection of the top dataset.

Technically, this boolean flag does not need to be retained. I hope it does not cause unwanted side effects across runs.