fix: make dataset files optional
Let's make all fields in object `.files` as well as `.files` field itself optional. The list of files does not make sense when using Auspice JSON as a dataset input.

This affects `pathogen.json` as well as `.meta.extensions.nextclade.pathogen.files`.

I had to add manual null checks in a few places for reference sequence and changelog files which were previously considered required.

If the reference sequence is not available from any of the possible sources (dataset file, Auspice JSON, `--input-ref` CLI arg or `&input-ref` URL param), there is no way for Nextclade to proceed, so an error is emitted in this case.