feat(web): take title, description and update date from Auspice JSON
When using Auspice JSON as an input dataset, if pathogen info is not present, let's also attempt to read `.meta.title` or `.meta.description` and use as a dataset name. And let's try to read `.meta.updated` as the updated date time of the dataset.

This allows for a prettier and more informative dataset info section when using Auspice JSON as an input dataset.